pyproject.toml
[build-system]
requires = ['poetry-core>=1.2.0', 'wheel',]
build-backend = 'poetry.core.masonry.api'


[tool.poetry]
name = 'complyscribe'
version = "0.13.0"
description = "complyscribe assists users in leveraging Compliance-Trestle in automated workflows of for OSCAL formatted compliance content management."

authors = ["Jennifer Power <jpower@redhat.com>",]

include = ['LICENSE']
exclude = ['tests/', 'docs/']
license = 'Apache-2.0'
readme = 'README.md'

repository = 'https://github.com/RedHatProductSecurity/complyscribe'


[tool.poetry.scripts]
complyscribe = "complyscribe.cli.root:root_cmd"

[tool.poetry.dependencies]
python = '^3.8.1'
gitpython = "^3.1.41"
compliance-trestle = "^3.8.1"
github3-py = "^4.0.1"
python-gitlab = "^4.2.0"
ruamel-yaml = "^0.18.13"
pydantic = "^2.0.0"
ssg = {git = "https://github.com/ComplianceasCode/content"}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
black = "^24.3.0"
mypy = "^1.11.0"
isort = "^5.13.2"
safety = "^3.2.8"
flake8-print = "^5.0.0"
pre-commit = "^3.4.0"
mkdocs-material = "^9.6.14"
markdown-include = "^0.8.1"
types-pyyaml = "^6.0.12.20240917"

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "^8.3.5"
pytest-cov = "^5.0.0"
pytest-skip-slow = "^0.0.5"
responses = "^0.25.7"

[tool.poetry.group.plugins]
optional = true

[tool.poetry.group.plugins.dependencies]
compliance-trestle-fedramp = "^0.4.0"

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    'tests/*',
]

[tool.pytest.ini_options]
minversion = '6.0'
addopts = """
--doctest-modules \
--cov=./ \
--cov-append \
--cov-report html:tests/reports/coverage-html \
--cov-report xml:tests/reports/coverage.xml \
--ignore=docs/
"""
testpaths = [
    'tests',
]

[tool.mypy]

plugins = [
  "pydantic.mypy"
]

follow_imports = "skip"
warn_redundant_casts = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = "github3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ruamel"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "responses"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ssg.*"
ignore_missing_imports = true
